+++ CineRating/Controllers/RealizadoresController.cs
@@ -51,7 +51,7 @@ namespace CineRating.Controllers {
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,Nome,DataNascimento,Biografia")] Realizadores realizadores, HttpPostedFileBase fileUploadImagem) {
+        public ActionResult Create([Bind(Include = "ID,Nome,DataNascimento,Biografia")] Realizadores realizadores, HttpPostedFileBase fileUploadImagem, string DataNasc) {
 
 
             //vars aux
@@ -68,7 +68,8 @@ namespace CineRating.Controllers {
                 return View(realizadores);
             }
 
-
+            DateTime dataNa = DateTime.Parse(DataNasc);
+            realizadores.DataNascimento = dataNa;
 
             if (ModelState.IsValid) {
                 db.Realizadores.Add(realizadores);
@@ -122,14 +123,17 @@ namespace CineRating.Controllers {
                 db.Entry(realizadores).State = EntityState.Modified;
                 db.SaveChanges();
 
-                if (fileUploadImagem != null) {
+                if ((fileUploadImagem != null) && (fileUploadImagem.ContentType.ToString() == "image/jpeg")) {
                     //guardar o nome da imagem na BD
                     fileUploadImagem.SaveAs(caminhoParaImagem);
                     System.IO.File.Delete(Path.Combine(Server.MapPath("~/imagens/atores"), oldName));
+                    return RedirectToAction("Details", "Realizadores", new { id = realizadores.ID });
                 }
-
-                return RedirectToAction("Index");
+                if (fileUploadImagem == null) {
+                    return RedirectToAction("Details", "Realizadores", new { id = realizadores.ID });
+                }
             }
+            ModelState.AddModelError("", "Não foi fornecida uma imagem ou o ficheiro inserido não é JPG");
             return View(realizadores);
         }
